Some final tweaks to Sonos S2 provider
